requirements.txt
dbt-core
python-dotenv
pandas
minio
pyarrow
polars
pandas
requests
openpyxl
duckdb
streamlit
altair
pytest
prefect
great_expectations==1.5.5